+++ src/containerapp/azext_containerapp/_validators.py
@@ -4,9 +4,13 @@
 # --------------------------------------------------------------------------------------------
 # pylint: disable=line-too-long, unused-argument
 
-from msrestazure.tools import is_valid_resource_id
 from knack.log import get_logger
 from urllib.parse import urlparse
+
+from azure.cli.core.azclierror import (ValidationError, InvalidArgumentValueError,
+                                       MutuallyExclusiveArgumentError, RequiredArgumentMissingError)
+from azure.cli.command_modules.containerapp._utils import is_registry_msi_system
+from azure.mgmt.core.tools import is_valid_resource_id
 
 from azure.cli.command_modules.containerapp._utils import is_registry_msi_system, safe_get
 from azure.cli.command_modules.containerapp._validators import _validate_revision_exists, _validate_replica_exists, \
@@ -68,7 +72,7 @@ def validate_runtime(runtime, enable_java_metrics, enable_java_agent):
 
 def validate_env_name_or_id(cmd, namespace):
     from azure.cli.core.commands.client_factory import get_subscription_id
-    from msrestazure.tools import resource_id, parse_resource_id
+    from azure.mgmt.core.tools import resource_id, parse_resource_id
 
     if not namespace.managed_env:
         return
@@ -111,7 +115,7 @@ def validate_env_name_or_id(cmd, namespace):
 
 def validate_env_name_or_id_for_up(cmd, namespace):
     from azure.cli.core.commands.client_factory import get_subscription_id
-    from msrestazure.tools import resource_id, parse_resource_id
+    from azure.mgmt.core.tools import resource_id, parse_resource_id
 
     if not namespace.environment:
         return
@@ -154,7 +158,7 @@ def validate_env_name_or_id_for_up(cmd, namespace):
 
 def validate_custom_location_name_or_id(cmd, namespace):
     from azure.cli.core.commands.client_factory import get_subscription_id
-    from msrestazure.tools import resource_id
+    from azure.mgmt.core.tools import resource_id
 
     if not namespace.custom_location or not namespace.resource_group_name:
         return
@@ -219,49 +223,3 @@ def validate_timeout_in_seconds(cmd, namespace):
     if timeout_in_seconds is not None:
         if timeout_in_seconds < 0 or timeout_in_seconds > 60:
             raise ValidationError("timeout in seconds must be in range [0, 60].")
-
-
-def validate_debug(cmd, namespace):
-    print("Validating...")
-    revision_already_set = bool(namespace.revision)
-    replica_already_set = bool(namespace.replica)
-    container_already_set = bool(namespace.container)
-    _set_debug_defaults(cmd, namespace)
-    if revision_already_set:
-        _validate_revision_exists(cmd, namespace)
-    if replica_already_set:
-        _validate_replica_exists(cmd, namespace)
-    if container_already_set:
-        _validate_container_exists(cmd, namespace)
-
-
-def _set_debug_defaults(cmd, namespace):
-    app = ContainerAppPreviewClient.show(cmd, namespace.resource_group_name, namespace.name)
-    if not app:
-        raise ResourceNotFoundError("Could not find a container app")
-    if not namespace.revision:
-        namespace.revision = app.get("properties", {}).get("latestRevisionName")
-        if not namespace.revision:
-            raise ResourceNotFoundError("Could not find a revision")
-    if not namespace.replica:
-        replicas = ContainerAppPreviewClient.list_replicas(
-            cmd=cmd,
-            resource_group_name=namespace.resource_group_name,
-            container_app_name=namespace.name,
-            revision_name=namespace.revision
-        )
-        if not replicas:
-            raise ResourceNotFoundError("Could not find a active replica")
-        namespace.replica = replicas[0]["name"]
-        if not namespace.container and replicas[0]["properties"]["containers"]:
-            namespace.container = replicas[0]["properties"]["containers"][0]["name"]
-    if not namespace.container:
-        revision = ContainerAppPreviewClient.show_revision(
-            cmd,
-            resource_group_name=namespace.resource_group_name,
-            container_app_name=namespace.name,
-            name=namespace.revision
-        )
-        revision_containers = safe_get(revision, "properties", "template", "containers")
-        if revision_containers:
-            namespace.container = revision_containers[0]["name"]