+++ src/serviceconnector-passwordless/HISTORY.rst
@@ -2,6 +2,18 @@
 
 Release History
 ===============
+3.1.0
+++++++
+* Add `az aks connection create`
+
+3.0.2
+++++++
+* Some improvements and security issue fixes.
+
+3.0.1
+++++++
+* Some improvements and security issue fixes.
+
 3.0.0
 ++++++
 * Add new param --new to override the existing database user and deprecate Postgres single server